services/user/requirements.txt
Flask
Flask_restful
flask_sqlalchemy
flask_jwt_extended
mysql-python
pyjwt
